requirements.txt
# 向量检索
faiss-cpu>=1.7.4

# 关键词检索
rank-bm25>=0.2.2
jieba>=0.42.1

# 嵌入模型（ModelScope GTE-base）
modelscope>=1.9.0

# 知识图谱（graphiti + KuZu）
graphiti-core[kuzu]>=0.28.0

# LLM 调用（阿里云 DashScope OpenAI 兼容接口）
openai>=1.0.0

# 环境变量与配置文件
python-dotenv>=1.0.0

# 科学计算
numpy>=1.24.0


# mcp
mcp>=1.6.0
httpx>=0.27.0
uvicorn>=0.27.0
starlette>=0.37.0


